Fix types.

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: jseijo-p <jseijo-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/19 10:11:37 by jseijo-p          #+#    #+#             */
-/*   Updated: 2022/04/20 10:53:54 by jseijo-p         ###   ########.fr       */
+/*   Updated: 2022/04/20 17:52:05 by jseijo-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@
 
 void	ft_bzero(void *s, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
